Update and rename kerningham.rs to kerninghan.rs
